+++ src/schemathesis/filters.py
@@ -16,8 +16,4 @@ def should_skip_endpoint(endpoint: str, pattern: Optional[Filter]) -> bool:
     if pattern is None:
         return False
     patterns = force_tuple(pattern)
-    return not any(is_match(endpoint, item) for item in patterns)
-
-
-def is_match(endpoint: str, pattern: str) -> bool:
-    return pattern in endpoint or bool(re.search(pattern, endpoint))
+    return not any(re.search(item, endpoint) for item in patterns)