+++ docs/changelog.rst
@@ -6,6 +6,9 @@ Changelog
 `Unreleased`_
 -------------
 
+`0.6.0`_ - 2019-09-24
+---------------------
+
 Added
 ~~~~~
 
@@ -106,7 +109,8 @@ Fixed
 
 - Initial public release
 
-.. _Unreleased: https://github.com/kiwicom/schemathesis/compare/v0.5.0...HEAD
+.. _Unreleased: https://github.com/kiwicom/schemathesis/compare/v0.6.0...HEAD
+.. _0.6.0: https://github.com/kiwicom/schemathesis/compare/v0.5.0...v0.6.0
 .. _0.5.0: https://github.com/kiwicom/schemathesis/compare/v0.4.1...v0.5.0
 .. _0.4.1: https://github.com/kiwicom/schemathesis/compare/v0.4.0...v0.4.1
 .. _0.4.0: https://github.com/kiwicom/schemathesis/compare/v0.3.0...v0.4.0